@@ -1,68 +1,67 @@
 {
-    "name": "a7-multi-comp",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "build-dev": "ng build --base-href / -c=dev",
-        "build-staging": "ng build --base-href / -c=staging",
-        "build-production": "ng build --base-href / -c=production",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "test-single-headless": "ng test --watch=false --browsers=PhantomJS --karma-config=src/karma.conf.phantom.js",
-        "test-sourcemaps": "ng test --source-map=false"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "~7.0.0",
-        "@angular/cdk": "^7.0.4",
-        "@angular/common": "~7.0.0",
-        "@angular/compiler": "~7.0.0",
-        "@angular/core": "~7.0.0",
-        "@angular/forms": "~7.0.0",
-        "@angular/http": "~7.0.0",
-        "@angular/material": "^7.0.4",
-        "@angular/platform-browser": "~7.0.0",
-        "@angular/platform-browser-dynamic": "~7.0.0",
-        "@angular/router": "~7.0.0",
-        "@auth0/angular-jwt": "^2.0.0",
-        "angular-highcharts": "^7.2.0",
-        "bootstrap": "^4.3.1",
-        "core-js": "^2.5.4",
-        "esri-loader": "^2.6.0",
-        "font-awesome": "^4.7.0",
-        "highcharts": "^7.1.1",
-        "karma-junit-reporter": "^1.2.0",
-        "ng-http-loader": "^2.3.0",
-        "ngx-popper": "^4.2.0",
-        "ngx-toastr": "^9.1.1",
-        "rxjs": "~6.3.3",
-        "stacktrace-js": "^2.0.0",
-        "web-animations-js": "^2.3.1",
-        "zone.js": "~0.8.26"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^0.13.9",
-        "@angular/cli": "^7.3.9",
-        "@angular/compiler-cli": "^7.2.15",
-        "@angular/language-service": "~7.0.0",
-        "@types/jasmine": "~2.8.8",
-        "@types/jasminewd2": "~2.0.3",
-        "@types/node": "~8.9.4",
-        "codelyzer": "~4.5.0",
-        "jasmine-core": "~2.99.1",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~3.0.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "karma-junit-reporter": "^1.2.0",
-        "protractor": "~5.4.0",
-        "ts-node": "~7.0.0",
-        "tslint": "~5.11.0",
-        "typescript": "~3.1.1"
-    }
+  "name": "a7-multi-comp",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "build-dev": "ng build --base-href / -c=dev",
+    "build-staging": "ng build --base-href / -c=staging",
+    "build-production": "ng build --base-href / -c=production",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "test-single-headless": "ng test --watch=false --browsers=PhantomJS --karma-config=src/karma.conf.phantom.js",
+    "test-sourcemaps": "ng test --source-map=false"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~8.2.5",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "~8.2.5",
+    "@angular/compiler": "~8.2.5",
+    "@angular/core": "~8.2.5",
+    "@angular/forms": "~8.2.5",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "~8.2.5",
+    "@angular/platform-browser-dynamic": "~8.2.5",
+    "@angular/router": "~8.2.5",
+    "@auth0/angular-jwt": "^3.0.0",
+    "angular-highcharts": "^7.2.0",
+    "bootstrap": "^4.3.1",
+    "core-js": "^2.6.9",
+    "esri-loader": "^2.10.0",
+    "font-awesome": "^4.7.0",
+    "highcharts": "^7.2.0",
+    "karma-junit-reporter": "^1.2.0",
+    "ng-http-loader": "^6.0.1",
+    "ngx-popper": "^4.2.0",
+    "ngx-toastr": "^9.2.0",
+    "rxjs": "~6.5.3",
+    "stacktrace-js": "^2.0.0",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "~0.9.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.803.3",
+    "@angular/cli": "^8.3.3",
+    "@angular/compiler-cli": "^8.2.5",
+    "@angular/language-service": "~8.2.5",
+    "@types/jasmine": "^2.8.16",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "^4.3.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "^2.0.6",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-junit-reporter": "^1.2.0",
+    "protractor": "^5.4.2",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.5.3"
+  }
 }
